+++ src/Ckb.Molecule/Type/ScriptSerializer.cs
@@ -8,7 +8,7 @@ namespace Ckb.Molecule.Type
         public ScriptSerializer(Script script)
             : base(new BaseSerializer[] {
                 new Byte32Serializer(script.CodeHash),
-                new ByteSerializer(script.HashType == "data" ? 0x0 : 0x1),
+                new ByteSerializer((byte)(script.HashType == "data" ? 0x0 : 0x1)),
                 new BytesSerializer(script.Args),
             })
         { }
